# Set is much faster than array for checking uniqueness