@aetherknight 2017-04-29
Looks good. I'll get it merged in and cut a new release in a moment.